simplify chinchilla usage at the cost of performance.
shouldComponentUpdate should always be true now. This way, every
component gets its latest model on every tree update, obviating the
need for `data.latest.findNode(this.props.model)` on modification. If
we could make components take the props but refuse to render, that
would be ideal. Maybe down the road.